+++ BTDB/KVDBLayer/ArtMemoryImplementation/ArtInMemoryKeyValueDBTransaction.cs
@@ -319,7 +319,7 @@ namespace BTDB.KVDBLayer
 
         public bool IsWriting()
         {
-            return _writing;
+            return _writing || _preapprovedWriting;
         }
 
         public ulong GetCommitUlong()