+++ BTDB/KVDBLayer/Implementation/KeyValueDBTransaction.cs
@@ -335,7 +335,7 @@ namespace BTDB.KVDBLayer
 
         public bool IsWriting()
         {
-            return _writing;
+            return _writing || _preapprovedWriting;
         }
 
         public ulong GetCommitUlong()