+++ BTDB/KVDBLayer/MemoryImplementation/InMemoryKeyValueDBTransaction.cs
@@ -294,7 +294,7 @@ namespace BTDB.KVDBLayer
 
         public bool IsWriting()
         {
-            return _writing;
+            return _writing || _preapprovedWriting;
         }
 
         public ulong GetCommitUlong()