+++ BTDB/ODBLayer/IType2NameRegistry.cs
@@ -6,6 +6,6 @@ namespace BTDB.ODBLayer
     {
         string RegisterType(Type type, string asName);
         Type FindTypeByName(string name);
-        string FindNameByType(Type type);
+        string? FindNameByType(Type type);
     }
 }